requirements.txt
cudatoolkit==10.2.89
pillow==9.0.1
scikit-learn==1.0.2
tqdm==4.63.1
pytorch-lightning==1.6.0